Move "peer/single".New to "peer".NewSingle
